+++ core/iwasm/libraries/debug-engine/handler.c
@@ -455,7 +455,7 @@ handle_threadstop_request(WASMGDBServer *server, char *payload)
     }
 
     tid = debug_inst->stopped_thread->handle;
-    status = debug_inst->stopped_thread->current_status->signal_flag;
+    status = (uint32)debug_inst->stopped_thread->current_status->signal_flag;
 
     wasm_debug_instance_set_cur_thread(debug_inst, tid);
